from TargetLanguageActions moved methods into importHelpers.js and tStudioHelpers.js
